Avoids overflow when checking number of missed blocks.

+++ integration/simulation/validate_chain.go
@@ -145,7 +145,7 @@ func checkBlockchainOfObscuroNode(t *testing.T, node *host.Node, minObscuroHeigh
 	l1Height := node.DB().GetCurrentBlockHead().Height
 
 	// check that the L1 view is consistent with the L1 network.
-	if maxEthereumHeight-l1Height > MAX_BLOCK_DELAY {
+	if int64(maxEthereumHeight)-int64(l1Height) > MAX_BLOCK_DELAY {
 		t.Errorf("Obscuro node %d fell behind %d blocks.", obscurocommon.ShortAddress(node.ID), maxEthereumHeight-l1Height)
 	}
